Added all examples
