+++ .github/workflows/examples-testing.yml
@@ -1,0 +1,24 @@
+name: Examples testing
+on:
+  push:
+  pull_request:
+    types: [opened, reopened, synchronize, ready_for_review]
+jobs:
+  test:
+    if: github.event.pull_request.draft == false
+    name: Examples testing ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v2
+      - name: Setup Node.js
+        uses: actions/setup-node@v1
+        with:
+          node-version: 14
+      - name: Build library
+        run: npm install && npm run build
+      - name: Test output
+        run: npm run test:examples
