+++ docs/advanced.md
@@ -53,23 +53,7 @@ The library uses 4 different logging levels:
 - `warn`: for warnings a user may need if the output is not as expected
 - `error`: for errors that occur in the library
 
-### Example usage
-Below is an example integration for adding a custom logger to the library:
-
-```ts
-import {ModelLoggingInterface, Logger} from '@asyncapi/modelina'; 
-const customLogger: ModelLoggingInterface = {
-    debug: (msg: string) => { console.log(msg) },
-    info: (msg: string) => { console.log(msg) },
-    warn: (msg: string) => { console.log(msg) },
-    error: (msg: string) => { console.log(msg) }
-};
-Logger.setLogger(customLogger);
-
-// Now use the library as normal without doing anything else. 
-// const generator = new TypeScriptGenerator({ modelType: 'interface' });
-// const interfaceModel = await generator.generate(...);
-```
+Check out this [example out for a live demonstration](../examples/custom-logging).
 
 ## Change the generated indentation type and size
 TODO
