+++ examples/asyncapi-from-parser/index.spec.ts
@@ -1,5 +1,5 @@
 import {generate} from './index';
-test('Should render typescript interface', async () => {
+test('Should be able to process AsyncAPI object from parser', async () => {
   const models = await generate();
   expect(models).not.toBeUndefined();
   expect(models).not.toHaveLength(0);
