+++ examples/json-schema-draft7-from-object/index.spec.ts
@@ -1,5 +1,5 @@
 import {generate} from './index';
-test('Should render typescript interface', async () => {
+test('Should be able to process JSON Schema draft 7 object', async () => {
   const models = await generate();
   expect(models).not.toBeUndefined();
   expect(models).not.toHaveLength(0);
