+++ package.json
@@ -70,8 +70,8 @@
     "docker:test": "npm run docker:build && docker run asyncapi/modelina npm run test",
     "docker:test:blackbox": "npm run docker:build && docker run asyncapi/modelina npm run test:blackbox",
     "test": "cross-env CI=true jest --coverage --testPathIgnorePatterns ./test/blackbox",
-    "test:snapshot": "npm run test -- --updateSnapshot",
     "test:blackbox": "cross-env CI=true jest ./test/blackbox",
+    "test:examples": "cross-env CI=true jest ./examples",
     "test:watch": "jest --watch",
     "docs": "npm run docs:markdown",
     "docs:markdown": "jsdoc2md lib/cjs/index.js -f lib/cjs/**/*.js > API.md",
